refactor into directory_listing_download, polish
- make dld a feature flag
- handle conditional compilation
- fix clippy errors
